uses hook to get roadtripId
